feat: support Extension<RequestId>
Looks like tower-request-id library could not be extracted due to
it's `RequestId` not being `Clone`. This is fine anyway, as i wanted
to embed a `Uuid` type for UserId, so RequestId can just use the same.

So this removes tower-request-id, implements the necessary Axum glue
in way of Service and Layer, adds the new Uuid related types.