feat: back to BASE64URL_NOPAD
After some experimentation, encoded length seemed to be the most
important factor. It may be present several times in a URL and
overall is the largest concern, even over double-click friendly.

Reverting back to Base64. Also exploring using NanoIDs to take
this to the extreme, but i don't think i want to give up sorted
values.